+++ src/db/knexfile.js
@@ -1,0 +1,52 @@
+import dotenv from 'dotenv'
+//Needed to specify multiple paths, because by default, config will look for a file called .env in the current working directory.
+//For migrations the cwd will be src/db, and generally for running the main app, the cwd will be /src
+dotenv.config({ path: ['../../.env', '../.env'] })
+
+console.log('host:');
+console.log(process.env.DEV_DATABASE_HOST);
+export const development = {
+  client: 'mysql2',
+  connection: {
+    host: process.env.DEV_DATABASE_HOST,
+    port: process.env.DEV_DATABASE_PORT,
+    user: process.env.DEV_DATABASE_USER,
+    password: process.env.DEV_DATABASE_PASSWORD,
+    database: process.env.DEV_DATABASE_NAME,
+  },
+  migrations: {
+    tableName: 'knex_migrations',
+  },
+};
+
+export const staging = {
+  client: 'postgresql',
+  connection: {
+    database: 'my_db',
+    user: 'username',
+    password: 'password',
+  },
+  pool: {
+    min: 2,
+    max: 10,
+  },
+  migrations: {
+    tableName: 'knex_migrations',
+  },
+};
+
+export const production = {
+  client: 'postgresql',
+  connection: {
+    database: 'my_db',
+    user: 'username',
+    password: 'password',
+  },
+  pool: {
+    min: 2,
+    max: 10,
+  },
+  migrations: {
+    tableName: 'knex_migrations',
+  },
+};
